+++ R/kobo_token.R
@@ -13,9 +13,11 @@
 #'
 #' @examples
 #' \dontrun{
-#' token <- kobo_setup(username = "cool_user_name",
-#'                     password = "xsww3@dddb",
-#'                     url = "https://kf.kobotoolbox.org/")
+#' if (require(askpass)) {
+#'  token <- kobo_setup(username = "cool_user_name",
+#'                      password = askpass::askpass(),
+#'                      url = "https://kf.kobotoolbox.org/")
+#' }
 #' token
 #' }
 #'