+++ tests/testthat/test-kobo_form.R
@@ -33,7 +33,6 @@ test_that("kobo_form has a version argument", {
                c("tbl_df", "tbl", "data.frame"))
  })
 
-
 test_that("kobo_form can load form with choices tab", {
   url <- Sys.getenv("KOBOTOOLBOX_PROD_URL")
   token <- Sys.getenv("KOBOTOOLBOX_PROD_TOKEN")